fix: separate web and node arweave initialization
