+++ src/web/utils/arweave.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2023 Fair protocol
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import Arweave from 'arweave/web';
+import { JWKInterface } from 'warp-contracts';
+import { NET_ARWEAVE_URL } from '../../common/utils/constants';
+import { ITag } from '../../common/types/arweave';
+
+const arweave = Arweave.init({
+  host: NET_ARWEAVE_URL.split('//')[1],
+  port: 443,
+  protocol: 'https',
+});
+
+export const jwkToAddress = async (jwk: JWKInterface) => arweave.wallets.jwkToAddress(jwk);
+
+export const getArBalance = async (address: string) => {
+  const winstonBalance = await arweave.wallets.getBalance(address);
+
+  return arweave.ar.winstonToAr(winstonBalance);
+};
+
+export const createTx = async (data: string, tags: ITag[]) => {
+  const tx = await arweave.createTransaction({ data });
+
+  tags.forEach(({ name, value }) => tx.addTag(name, value));
+
+  return tx;
+};
